+++ src/lib/py/subcommands.py
@@ -1,0 +1,5 @@
+import Convert2MzML
+
+SUBCOMMANDS = [
+    Convert2MzML.Convert2MzML,
+]